add eslint-plugin-clean-styled-components and cleanup components
