+++ .eslintrc.js
@@ -12,7 +12,9 @@ module.exports = {
       '@typescript-eslint/no-var-requires': 'off',
     },
   },
+  plugins: ['clean-styled-components'],
   rules: {
+    'clean-styled-components/single-component-per-file': 'error',
     'no-underscore-dangle': ['error', { allow: ['__NEXT_DATA__'] }],
     'import/no-extraneous-dependencies': [
       'error',
