+++ components/footer.tsx
@@ -4,6 +4,7 @@ import Link from 'next/link';
 import styled from 'styled-components';
 import { name } from '~/utils/authorInfo';
 
+// eslint-disable-next-line clean-styled-components/single-component-per-file
 const Footer = styled.footer`
   height: 8rem;
   background: ${({ theme }) => theme.background};
@@ -11,31 +12,21 @@ const Footer = styled.footer`
   display: flex;
   justify-content: center;
   align-items: center;
-`;
 
-const StyledLink = styled.a`
-  color: currentColor;
-  text-decoration: none;
+  a {
+    color: currentColor;
+    text-decoration: none;
+  }
 `;
 
 const FooterWrap = () => {
-  const router = useRouter();
-  const isHome = router.pathname === '/';
+  const { query } = useRouter();
   return (
     <Footer>
-      <Link href={{ pathname: '/', query: router.query }} passHref>
-        <StyledLink
-          rel="home"
-          aria-label="go home"
-          onClick={event => {
-            if (isHome) {
-              event.preventDefault();
-              window.scrollTo({ top: 0, behavior: 'smooth' });
-            }
-          }}
-        >
+      <Link href={{ pathname: '/', query }}>
+        <a rel="home" aria-label="go home">
           &copy; {new Date().getFullYear()} {name}
-        </StyledLink>
+        </a>
       </Link>
     </Footer>
   );
