+++ package.json
@@ -59,6 +59,7 @@
     "@types/is-absolute-url": "3.0.0",
     "@types/jest": "24.0.13",
     "@types/jest-in-case": "1.0.1",
+    "@types/node": "12.0.2",
     "@types/nprogress": "0.0.29",
     "@types/prettier": "1.16.4",
     "@types/react": "16.8.18",
@@ -80,6 +81,7 @@
     "eslint-config-prettier": "4.3.0",
     "eslint-import-resolver-babel-plugin-root-import": "1.1.1",
     "eslint-import-resolver-typescript": "1.1.1",
+    "eslint-plugin-clean-styled-components": "0.0.2",
     "eslint-plugin-compat": "3.1.1",
     "eslint-plugin-import": "2.17.3",
     "eslint-plugin-jest": "22.6.4",
