+++ types/window.d.ts
@@ -3,3 +3,9 @@ interface Window {
   ReactIntlLocaleData: any;
   __NEXT_DATA__: any;
 }
+
+declare namespace NodeJS {
+  interface Process {
+    browser: boolean;
+  }
+}
